Rewrite on fsevents: macOS specific
This is a platform-specific library but offers a much more useful event
stream; eliminmates a lot of the resampling cruft from the prior
fsnotify solution.

This looks like it *just works*.

Known unknown: are there event types that should be filtered out by
default? The event flag pattern is a little unclear; should look for
examples.

Revisiting the terminal display.